move logic for loc api ingestion
